+++ frontend/src/Components/Admin/Grupos.js
@@ -29,11 +29,21 @@ class AdminGroups extends React.Component {
         let username=document.getElementById('editUsernameInput').value;
         let user={...this.state.actualGroup};
         user.nombre=username;
+        let tok = localStorage.getItem('token')
+        if(tok){
+        axios.defaults.headers.common['Authorization'] = 
+                                'Bearer ' + localStorage.getItem('token').substring(1, localStorage.getItem('token').length - 1);
         axios.put('http://localhost:3001/Fansubs/'+user.id,user);
+        }
     }
 
     deleteUsuario=(idUser)=>{
+        let tok = localStorage.getItem('token')
+        if(tok){
+        axios.defaults.headers.common['Authorization'] = 
+                                'Bearer ' + localStorage.getItem('token').substring(1, localStorage.getItem('token').length - 1);
         axios.delete('http://localhost:3001/Fansubs/'+idUser);
+        }
     }
 
     rendGrupos=()=>{
